New tab created using admin account
