--- teams/TailwindTraders-M365/tabs/.env
@@ -1,3 +1,0 @@
-HTTPS=true
-BROWSER=none
-
