+++ teams/TailwindTraders-M365/tabs/Learn Together Tab/.env
@@ -1,0 +1,3 @@
+HTTPS=true
+BROWSER=none
+
