+++ teams/TailwindTraders-M365/tabs/Learn Together Tab/package.json
@@ -2,16 +2,16 @@
     "name": "microsoft-teams-ext",
     "version": "0.1.0",
     "msteams": {
-        "teamsAppId": "6ccd97e1-6b33-4864-984f-22de1c5f1e04"
+        "teamsAppId": "f2569384-bd2b-459d-9bca-66c5c4d0e920"
     },
     "private": true,
     "dependencies": {
+        "@microsoft/teams-js": "^1.6.0",
         "@fluentui/react-northstar": "^0.49.0",
-        "@microsoft/teams-js": "^1.8.0",
         "react": "^16.13.1",
         "react-dom": "^16.13.1",
-        "react-router-dom": "^5.1.2",
-        "react-scripts": "3.4.1"
+        "react-scripts": "3.4.1",
+        "react-router-dom": "^5.1.2"
     },
     "scripts": {
         "start": "react-scripts start",
