--- teams/TailwindTraders-M365/tabs/src/index.js
@@ -1,14 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './components/App';
-import { Provider, themes } from '@fluentui/react-northstar' //https://fluentsite.z22.web.core.windows.net/quick-start
-
-ReactDOM.render(
-    <Provider theme={themes.teams}>
-        <App />
-    </Provider>, document.getElementById('root')
-);
